Move waitForNewTurn from t3code/index.ts to t3code/client.ts
Relocates the `waitForNewTurn` polling utility to the lower-level
`client.ts` module where it belongs alongside other t3code client
primitives. Updates the import in `index.ts` and the test file.

`pollUntilDone` in the orchestration runner uses a broader, multi-agent
phase-polling mechanism (peer-sync + turn lifecycle state machine +
event-driven wakeup) that does not overlap with the single-thread,
snapshot-only approach of `waitForNewTurn`, so no runner changes are
needed.
